remorganize sys mod
