+++ crates/containerd-shim-wasm/src/sys/unix/mod.rs
@@ -1,0 +1,2 @@
+pub mod metrics;
+pub mod networking;
