+++ crates/containerd-shim-wasm/src/sys/windows/networking.rs
@@ -1,0 +1,4 @@
+pub fn setup_namespaces(spec: &oci_spec::runtime::Spec) -> anyhow::Result<()> {
+    // noop for now
+    Ok(())
+}
